requirements.txt
Flask
flask-cors
pandas
numpy
scikit-learn
scipy
gunicorn
psycopg2-binary
python-dotenv
sentence-transformers